% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{add_chunk_group}
\alias{add_chunk_group}
\title{Addition of chunk groups}
\usage{
add_chunk_group(data, rows_per_chunk)
}
\arguments{
\item{data}{df to partition into chunks}

\item{rows_per_chunk}{size of a chunk}
}
\description{
partition data into chunks of a fixed number of rows in order to calculate
  aggregated features per chunk
}
