% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\alias{aggregate_e4_data}
\alias{aggregate_embrace_plus_data}
\title{Aggregate data into 1min timestamps}
\usage{
aggregate_data(x)

aggregate_e4_data(x)

aggregate_embrace_plus_data(x)
}
\arguments{
\item{x}{An object read by \code{\link{read_e4}} or \code{\link{read_embrace_plus}}.}
}
\description{
Aggregate data into 1min timestamps

Aggregate E4 data into 1min timesteps

Aggregate Embrace Plus data into 1min timesteps
}
