% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_e4_data}
\alias{aggregate_e4_data}
\title{Aggregate E4 data into 1min timesteps}
\usage{
aggregate_e4_data(x)
}
\arguments{
\item{x}{An object read by \code{\link{read_e4}}.}
}
\description{
Aggregate E4 data into 1min timesteps
}
