% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_embrace_plus_data}
\alias{aggregate_embrace_plus_data}
\title{Aggregate Embrace Plus data into 1min timesteps}
\usage{
aggregate_embrace_plus_data(x)
}
\arguments{
\item{x}{An object read by \code{\link{read_embrace_plus}} and binded
with \code{\link{rbind_embrace_plus}}.}
}
\description{
Aggregate Embrace Plus data into 1min timesteps
}
