% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_time}
\alias{as_time}
\title{as_time}
\usage{
as_time(x)
}
\arguments{
\item{x}{takes a unixtime and converts to as.POSIXct}

\item{origin}{defaults to 1970-1-1}

\item{tz}{defaults to "UTC"}
}
\description{
Converts Unix time to as.POSIXct
}
