% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_timeseries.R
\name{as_timeseries}
\alias{as_timeseries}
\title{Convert an E4 data stream to a timeseries}
\usage{
as_timeseries(data, index = 2, name_col = "V1")
}
\arguments{
\item{data}{A dataframe, subelements of list as output by read_e4 function}

\item{index}{Which column (integer) to use as the data in the timeseries. Default: 2.}

\item{name_col}{Column name to give to the timeseries data.}
}
\description{
Creates an xts object indexed by time
}
