% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_analysis.R
\name{batch_analysis}
\alias{batch_analysis}
\title{Read and process all ZIP files in a directory}
\usage{
batch_analysis(path_in = choose.dir(), path_out = ".")
}
\description{
Read and process all ZIP files in a directory
}
