% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibi_analysis_rmssd_time_period.R
\name{calculate_RMSSF}
\alias{calculate_RMSSF}
\title{RMSSD calculation}
\usage{
calculate_RMSSF(IBIdata)
}
\arguments{
\item{IBIdata}{Uses the IBI data frame as created by \code{\link{read_e4}}}
}
\description{
Calculation of RMSSD over 1 minute time periods for plotting
}
