% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_e4_data_intervals.R
\name{char_clock_systime}
\alias{char_clock_systime}
\title{Force character datetime variable ("yyyy-mm-dd hh:mm:ss") to system timezone}
\usage{
char_clock_systime(time)
}
\arguments{
\item{time}{Datetime variable ("yyyy-mm-dd hh:mm:ss")}
}
\description{
Force character datetime variable ("yyyy-mm-dd hh:mm:ss") to system timezone
}
