% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_derivative_features}
\alias{compute_derivative_features}
\title{Compute derivatives}
\usage{
compute_derivative_features(derivative, feature_name)
}
\arguments{
\item{derivative}{first dreivative}

\item{feature_name}{what is the name of the feature}
}
\description{
Compute derivatives
}
