% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_features2}
\alias{compute_features2}
\title{Features computation}
\usage{
compute_features2(data)
}
\arguments{
\item{data}{df with eda, filtered eda and timestamp columns}
}
\description{
Compute features for SVM
}
