% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_wavelet_coefficients}
\alias{compute_wavelet_coefficients}
\title{Wavelet coefficients}
\usage{
compute_wavelet_coefficients(data)
}
\arguments{
\item{data}{data with an EDA element}
}
\description{
Compute wavelet coefficients.
}
