% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_wavelet_features}
\alias{compute_wavelet_features}
\title{Compute wavelet features}
\usage{
compute_wavelet_features(wavelet_coeffcients)
}
\arguments{
\item{wavelet_coeffcients}{use the wavelet coefficients !!variable name}
}
\description{
compute wavelet features
}
