% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_freq_list}
\alias{create_empty_freq_list}
\title{Create an Empty Frequency List}
\usage{
create_empty_freq_list()
}
\value{
A named list with frequency measurement fields, mostly set to NA.
}
\description{
This function generates a list named 'freq' with predefined fields for frequency measurements. 
Most fields are initialized with NA values, except for 'type' which is set to "fourier".
}
\examples{
empty_freq_list <- create_empty_freq_list()
str(empty_freq_list)
}
