% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_time_list}
\alias{create_empty_time_list}
\title{Create an Empty Time List}
\usage{
create_empty_time_list()
}
\value{
A named list with all fields set to NA.
}
\description{
This function generates a list named 'time' with predefined fields, all set to NA.
The fields included are related to time-based measurements.
}
\examples{
empty_time_list <- create_empty_time_list()
str(empty_time_list)
}
