% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e4_data.R
\docType{data}
\name{e4_data}
\alias{e4_data}
\title{Small example dataset for e4}
\format{
An object of class \code{"e4data"}
}
\usage{
e4_data
}
\description{
Small example dataset for e4
}
\examples{
data(e4_data)
}
\keyword{datasets}
