% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e4_timeseries_plot.R
\name{e4_timeseries_plot}
\alias{e4_timeseries_plot}
\title{Timeseries plot of E4 data}
\usage{
e4_timeseries_plot(
  data,
  main_title = "",
  calendar_data = NULL,
  series_options = NULL
)
}
\arguments{
\item{data}{List of e4 data.}

\item{main_title}{Main title of the plot, defaults to empty string.}

\item{calendar_data}{Name to give to the timeseries data.}

\item{average_lines}{Do you want to add averages to the plot}

\item{yaxis_ranges}{What Range do you want to give the y axis?}

\item{colours}{The colours of the graph, defaults to blue, orange, red, green}
}
\description{
Creates 4 graphs of the E4 data, as used in the Empatica dashboard.
}
