% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_e4data_datetime.R
\name{filter_datetime}
\alias{filter_datetime}
\alias{filter_e4data_datetime}
\alias{filter_embrace_plus_data_timestamp}
\title{Filter data according to a datetime start and end}
\usage{
filter_datetime(data, start, end)

filter_e4data_datetime(data, start, end)

filter_embrace_plus_data_timestamp(data, start, end)
}
\arguments{
\item{data}{Object read with \code{\link{read_e4}} or \code{\link{read_embrace_plus}}}

\item{start}{Start Datetime (posixct)}

\item{end}{End Datetime (posixct)}
}
\description{
Filter data according to a datetime start and end

Filter all four datasets for a Datetime start + end

Filter all datasets for a Datetime start + end
}
