% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_e4data_datetime.R
\name{filter_e4data_datetime}
\alias{filter_e4data_datetime}
\title{Filter all four datasets for a Datetime start + end}
\usage{
filter_e4data_datetime(data, start, end)
}
\arguments{
\item{data}{Object read with \code{\ink{read_e4}}}

\item{start}{End Datetime (posixct)}
}
\description{
Filter all four datasets for a Datetime start + end
}
