% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{find_peaks}
\alias{find_peaks}
\title{This function finds the peaks of an EDA signal and adds basic properties to the datafile.
Also, peak_end is assumed to be no later than the start of the next peak. (Is this okay??)}
\usage{
find_peaks(
  data,
  offset,
  start_WT,
  end_WT,
  thres = 0,
  sampleRate = getOption("SAMPLE_RATE", 8)
)
}
\arguments{
\item{data:}{DataFrame with EDA as one of the columns and indexed by a datetimeIndex}

\item{offset:}{the number of rising samples and falling samples after a peak needed to be counted as a peak}

\item{start_WT:}{maximum number of seconds before the apex of a peak that is the "start" of the peak}

\item{end_WT:}{maximum number of seconds after the apex of a peak that is the "rec.t/2" of the peak, 50% of amp}

\item{thres:}{the minimum microsecond change required to register as a peak, defaults as 0 (i.e. all peaks count)}

\item{sampleRate:}{number of samples per second, default=8}
}
\value{
data$peaks:               list of binary, 1 if apex of SCR

data$peak_start:          list of binary, 1 if start of SCR

data$peak_start_times:    list of strings, if this index is the apex of an SCR, it contains datetime of start of peak

data$peak_end:            list of binary, 1 if rec.t/2 of SCR

data$peak_end_times:      list of strings, if this index is the apex of an SCR, it contains datetime of rec.t/2

data$amplitude:           list of floats,  value of EDA at apex - value of EDA at start

data$max_deriv:           list of floats, max derivative within 1 second of apex of SCR
}
\description{
This function finds the peaks of an EDA signal and adds basic properties to the datafile.
Also, peak_end is assumed to be no later than the start of the next peak. (Is this okay??)
}
