% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_apex}
\alias{get_apex}
\title{Get the apex of the electrodermal activity signal}
\usage{
get_apex(eda_deriv, offset)
}
\arguments{
\item{eda_deriv}{uses the eda derivative to find the apex}

\item{offset}{lookback}

\item{i}{which peak}

\item{peak_sign}{positive or negative}
}
\description{
finds the apex of the eda signal in a sliding window

finds the drop of a peak in a sliding window
}
\details{



}
