% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_decay_time}
\alias{get_decay_time}
\title{Get the decay times of the peaks}
\usage{
get_decay_time(data, i_apex_with_decay)
}
\arguments{
\item{data}{vector?}

\item{i_apex_with_decay}{get apex and decay time}
}
\description{
Get the decay times of the peaks
}
\details{

}
