% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_eda_deriv}
\alias{get_eda_deriv}
\title{Get the derivative of the electrodermal activity signal}
\usage{
get_eda_deriv(eda)
}
\arguments{
\item{eda}{uses the eda vector to find the first derivatives}
}
\description{
finds the first derivatives of the eda signal
}
