% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_half_amp}
\alias{get_half_amp}
\title{Half peak amp}
\usage{
get_half_amp(data, i)
}
\arguments{
\item{data}{df with peak info}

\item{i}{apex index}
}
\description{
Get the amplitude value halfway between peak start and apex
}
