% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_i_apex_with_decay}
\alias{get_i_apex_with_decay}
\title{Get the apex to calculate decay times of the peaks}
\usage{
get_i_apex_with_decay(data)
}
\arguments{
\item{data}{vector?}
}
\description{
Get the apex to calculate decay times of the peaks
}
\details{

}
