% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_i_end_per_apex}
\alias{get_i_end_per_apex}
\title{Peak end}
\usage{
get_i_end_per_apex(i, i_max_peak_end, max_lookahead)
}
\arguments{
\item{i_max_peak_end}{beginning of next peak or end of signal}

\item{max_lookahead}{max distance from apex to search for end}

\item{data}{df with peak info}
}
\description{
Get the end of the peak for each individual peak with some restrictions
}
