% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_i_half_rise}
\alias{get_i_half_rise}
\title{Half rise index}
\usage{
get_i_half_rise(i_peak_start, i_apex)
}
\arguments{
\item{i_peak_start}{peak start index}

\item{i_apex}{apex index}
}
\description{
Get the index of the half rise point for a single peak
}
