% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{get_kernel}
\alias{get_kernel}
\title{SVM kernel}
\usage{
get_kernel(kernel_transformation, sigma, columns)
}
\arguments{
\item{kernel_transformation}{Data matrix used to transform EDA features
into kernel values}

\item{sigma}{The inverse kernel width used by the kernel}

\item{columns}{Features computed from EDA signal}
}
\description{
Generate kernel needed for SVM
}
