% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_max_deriv}
\alias{get_max_deriv}
\title{Get local maximum?}
\usage{
get_max_deriv(data, eda_deriv, sample_rate)
}
\arguments{
\item{data}{vector?}

\item{eda_deriv}{uses the derivative to find a local maximum}

\item{sample_rate}{sample rate of the signal}
}
\description{
What is the local maximum of the peaks?
}
