% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_peak_end_times}
\alias{get_peak_end_times}
\title{Peak end times}
\usage{
get_peak_end_times(data)
}
\arguments{
\item{data}{df with peak info}
}
\description{
Get the end timstamp of the peaks
}
