% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_peak_start}
\alias{get_peak_start}
\title{Start of peaks}
\usage{
get_peak_start(data, sample_rate)
}
\arguments{
\item{data}{df with peak info}

\item{sample_rate}{sample rate of the signal}
}
\description{
Provide info for each measurement whether it is the start of a peak (0 or 1)
}
