% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_rise_events_for_apex}
\alias{get_rise_events_for_apex}
\title{Number of rises before apex}
\usage{
get_rise_events_for_apex(i, eda_deriv, max_lookback)
}
\arguments{
\item{i}{index of apex}

\item{eda_deriv}{first derivative of signal}

\item{max_lookback}{maximum lookback}
}
\description{
Get the number of measurement with a rise before a single apex
}
