% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_rise_time}
\alias{get_rise_time}
\title{Calculates the rise time of a peak}
\usage{
get_rise_time(eda_deriv, apices, sample_rate, start_WT)
}
\arguments{
\item{eda_deriv}{uses the first derivative of a peak}

\item{apices}{which apices are there}

\item{sample_rate}{sample rate of the signal}

\item{start_WT}{start of the wavelet 
Number of rises before peak}

\item{i}{which peak}

\item{max_lookback}{maximum lookback}
}
\description{
Calculates the rise time of a peak

Get the number of rises before peak
}
\details{



}
