% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_SCR_width}
\alias{get_SCR_width}
\title{Get the width of the peaks}
\usage{
get_SCR_width(data, i_apex_with_decay)
}
\arguments{
\item{data}{vector?}

\item{i_apex_with_decay}{get apex and decay time}
}
\description{
Get the width of the peaks
}
