% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{max_per_n}
\alias{max_per_n}
\title{Max value per segment of length n}
\usage{
max_per_n(values, n, output_length)
}
\arguments{
\item{values}{array of numbers}

\item{n}{length of each segment}

\item{output_length}{argument to adjust for final segment not being full}
}
\description{
Give the maximum value of a vector of values per segment of length n.
}
