% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\docType{data}
\name{multiclass_classifier_config}
\alias{multiclass_classifier_config}
\title{Configuration of the SVM algorithm for ternary classification}
\format{
An object of class \code{list} of length 4.
}
\usage{
multiclass_classifier_config
}
\description{
Configuration of the SVM algorithm for ternary classification
}
\references{
\url{https://eda-explorer.media.mit.edu/}
}
\author{
Sara Taylor \email{sataylor@mit.edu}
}
\keyword{datasets}
