% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peak_has_drop}
\alias{peak_has_drop}
\title{Which peak has a drop?}
\usage{
peak_has_drop(i, peak_sign, offset)
}
\arguments{
\item{i}{apex index}

\item{peak_sign}{upward of downward movement of signal}

\item{offset}{minimum number of downward measurements after the apex,
in order to be considered a peak}
}
\description{
finds the drop of a peak within an optional time window
}
