% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{plot_artifacts}
\alias{plot_artifacts}
\title{Artifact plots}
\usage{
plot_artifacts(labels, eda_data)
}
\arguments{
\item{labels}{labels with artifact classification}

\item{eda_data}{data upon which the labels are plotted}
}
\description{
Plot artifacts after eda_data is classified
}
