% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{predict_binary_classifier}
\alias{predict_binary_classifier}
\title{Predict binary classifier}
\usage{
predict_binary_classifier(data, binary_classifier_config)
}
\arguments{
\item{data}{vector?}

\item{binary_classifier_config}{what config?}
}
\value{
binary_classifier_config Selects the binary classifier
}
\description{
Predict binary classifier
}
