% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{predict_multiclass_classifier}
\alias{predict_multiclass_classifier}
\title{Ternary classifiers}
\usage{
predict_multiclass_classifier(data)
}
\arguments{
\item{data}{features from EDA signal}
}
\description{
Generate classifiers (artifact, unclear, no artifact)
}
