% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepend_time_column}
\alias{prepend_time_column}
\title{prepend_time_column}
\usage{
prepend_time_column(data, timestart, hertz, tz = Sys.timezone())
}
\arguments{
\item{data}{dataframe}

\item{timestart}{the start of the reording}

\item{hertz}{hertz in which the E4 data was recorded}

\item{tz}{The timezone, defaults to user timezone}
}
\description{
Comlumn binds a time_coumn to the dataframe
}
