% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e4_methods.R
\name{print.e4data}
\alias{print.e4data}
\title{Show class of object}
\usage{
\method{print}{e4data}(x, ...)
}
\arguments{
\item{x}{An e4 data list}

\item{\dots}{Further arguments currently ignored.}
}
\description{
Returns 'object of class'
}
