% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_data.R
\name{rbind_data}
\alias{rbind_data}
\title{Row-bind datasets}
\usage{
rbind_data(data)
}
\arguments{
\item{data}{List object containing data.frames}
}
\description{
Row-bind datasets
}
\details{
Serves as an internal function that is called by
 `rbind_e4`, `rbind_embrace_plus` and `rbind_nowatch`. It takes a list 
 of data.frames and row-binds them together.
}
\keyword{internal}
