% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_data.R
\name{rbind_e4}
\alias{rbind_e4}
\title{Row-bind E4 datasets}
\usage{
rbind_e4(data)
}
\arguments{
\item{data}{An object read in by `\code{\link{read_e4}}}
}
\description{
This function takes a list of E4 datasets and row-binds them
together.
}
