% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_data.R
\name{rbind_embrace_plus}
\alias{rbind_embrace_plus}
\title{Row-bind Embrace Plus datasets}
\usage{
rbind_embrace_plus(data)
}
\arguments{
\item{data}{An object read in by \code{\link{read_embrace_plus}}}
}
\description{
This function takes a list of Embrace Plus datasets and row-binds 
them together.
}
