% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_data.R
\name{rbind_nowatch}
\alias{rbind_nowatch}
\title{Row-bind NOWATCH datasets}
\usage{
rbind_nowatch(data)
}
\arguments{
\item{data}{An object read in by \code{\link{read_nowatch}}}
}
\description{
This function takes a list of NOWATCH datasets and row-binds them
together.
}
