% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_process_e4.R,
%   R/read_and_process_embrace_plus.R
\name{read_and_process_e4}
\alias{read_and_process_e4}
\alias{process_e4}
\alias{process_embrace_plus}
\title{Read, process and feature extraction of E4 data}
\usage{
read_and_process_e4(zipfile, tz = Sys.timezone())

process_e4(data)

process_embrace_plus(data)
}
\arguments{
\item{zipfile}{zip file with e4 data to be read}

\item{tz}{timezone where data were recorded (default system timezone)}

\item{data}{object from read_e4 function}
}
\value{
An object with processed data and analyses, object of class 'e4_analysis'.
}
\description{
Reads the raw ZIP file using `read_e4`,
  performs analyses with `ibi_analysis` and `eda_analysis`.
}
