% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_process_embrace_plus.R
\name{read_and_process_embrace_plus}
\alias{read_and_process_embrace_plus}
\title{Read, process and feature extraction of Embrace Plus data}
\usage{
read_and_process_embrace_plus(
  zipfile = NULL,
  folder = NULL,
  type = "raw",
  tz = Sys.timezone()
)
}
\arguments{
\item{zipfile}{zip file with embrace plus data to be read}

\item{folder}{A folder with the unzipped files. If this is provided, the zipfile is not used.}

\item{type}{The type of data contained in the zip file. Either "raw" or "aggregated".}

\item{tz}{timezone where data were recorded (default system timezone)}
}
\value{
An object with processed data and analyses, object of class 'embrace_plus_analysis'.
}
\description{
Reads the raw ZIP file using `read_embrace_plus`,
  performs analyses with `eda_analysis`.
}
