% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_e4.R
\name{read_e4}
\alias{read_e4}
\title{Read E4 data}
\usage{
read_e4(zipfile = NULL, tz = Sys.timezone())
}
\arguments{
\item{zipfile}{A zip file as exported by the instrument}

\item{tz}{The timezone used by the instrument (defaults to user timezone).}
}
\description{
Reads in E4 data as a list (with EDA, HR, Temp, ACC, BVP, IBI as dataframes), and prepends timecolumns
}
