% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_embrace_plus.R
\name{read_embrace_plus}
\alias{read_embrace_plus}
\title{Read Embrace Plus data}
\usage{
read_embrace_plus(zipfile)
}
\arguments{
\item{zipfile}{A zip file as exported by the instrument}
}
\description{
Reads in Embrace Plus data as a list (with EDA, HR, Temp, ACC, BVP, IBI as dataframes), and prepends timecolumns
}
\details{
This function reads in a zipfile as exported by Embrace Plus. Then it extracts the zipfiles in a temporary folder
and unzips them in the same temporary folder.

The unzipped files are avro files, and are read in with using `sparklyr`, which sets up a local Spark cluster.

The function returns an object of class "embrace_plus_data" with a prepended datetime columns.
The object contains a list with dataframes from the physiological signals.
}
\examples{
library(wearables)
# read_embrace_plus("yourpathtohezipfile.zip")
}
