% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nowatch.R
\name{read_nowatch}
\alias{read_nowatch}
\title{Read Nowatch data}
\usage{
read_nowatch(zipfile, tz = Sys.timezone())
}
\arguments{
\item{zipfile}{A zip file as exported by the instrument. Only aggregated data supported.}

\item{tz}{The timezone used by the instrument (defaults to user timezone).}

\item{type}{The type of data contained in the zip file.}
}
\description{
Reads in Nowatch data as a list, and prepends timecolumns
}
\details{
This function reads in a zipfile as exported by Nowatch. Then it extracts the zipfiles in a temporary folder
and unzips them in the same temporary folder.

The unzipped files are csv files.

The function returns an object of class "nowatch_data" with a prepended datetime columns.
The object contains a list with dataframes from the physiological signals.
}
\examples{
\dontrun{
 library(wearables)
 read_nowatch("yourpathtohezipfile.zip")
}
}
