% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{remove_small_peaks}
\alias{remove_small_peaks}
\title{Small peaks removal}
\usage{
remove_small_peaks(data, thres = 0)
}
\arguments{
\item{data}{df with info on peaks}

\item{thres}{threshold of amplitude difference in order to be removed
(default 0 means no removals)}
}
\description{
Remove peaks with a small rise from start to apex are removed
}
