% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{split_in_chunks}
\alias{split_in_chunks}
\title{Chunk splitting}
\usage{
split_in_chunks(data, rows_per_chunk)
}
\arguments{
\item{data}{vector of values}

\item{rows_per_chunk}{number of values bundled per chunk}
}
\description{
Split into chunks of equal length.
}
