% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{upsample_data_to_8Hz}
\alias{upsample_data_to_8Hz}
\title{Upsample EDA data to 8 Hz}
\usage{
upsample_data_to_8Hz(eda_data)
}
\arguments{
\item{eda_data}{Data read with \code{\link{read_e4}}}
}
\description{
Upsample EDA data to 8 Hz
}
