% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{write_peak_features}
\alias{write_peak_features}
\title{Write peak features}
\usage{
write_peak_features(data_with_peaks, outfile)
}
\arguments{
\item{data_with_peaks}{vector with peaks}

\item{outfile}{where should it be written to}
}
\description{
Write peak features
}
