% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_process_e4.R
\name{write_processed_e4}
\alias{write_processed_e4}
\title{Write CSV files of the output}
\usage{
write_processed_e4(obj, out_path = ".")
}
\arguments{
\item{obj}{e4 analysis object}

\item{out_path}{output folder}
}
\description{
Slow!
}
