% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{choose_between_classes}
\alias{choose_between_classes}
\title{Which classifier?}
\usage{
choose_between_classes(class_a, class_b, kernels)
}
\arguments{
\item{class_a}{vector?}

\item{class_b}{vector?}

\item{kernels}{kernel?}
}
\description{
Which classifier?
}
