% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_amplitude_features}
\alias{compute_amplitude_features}
\title{Compute amplitude features}
\usage{
compute_amplitude_features(data)
}
\arguments{
\item{data}{vector?}
}
\description{
Compute amplitude features
}
