% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_derivative_features}
\alias{compute_derivative_features}
\title{Derivative features}
\usage{
compute_derivative_features(derivative, feature_name)
}
\arguments{
\item{derivative}{vector of derivatives}

\item{feature_name}{name of feature}
}
\description{
Compute derivative features.
}
