% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_wavelet_coefficients}
\alias{compute_wavelet_coefficients}
\title{Compute wavelet coefficients}
\usage{
compute_wavelet_coefficients(data)
}
\arguments{
\item{data}{vector?}
}
\description{
compute wavelet coefficients
}
