% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{compute_wavelet_decomposition}
\alias{compute_wavelet_decomposition}
\title{Compute wavelet decomposition}
\usage{
compute_wavelet_decomposition(data)
}
\arguments{
\item{data}{vector?}
}
\description{
compute wavelet decomposition
}
