% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_proces_e4.R
\name{create_e4_output_folder}
\alias{create_e4_output_folder}
\title{Create output folder for E4 analysis results}
\usage{
create_e4_output_folder(obj, out_path = ".")
}
\description{
Create output folder for E4 analysis results
}
