% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_decay_time}
\alias{get_decay_time}
\title{Decay time}
\usage{
get_decay_time(data, i_apex_with_decay)
}
\arguments{
\item{data}{df with peak info}

\item{i_apex_with_decay}{indexes of relevant peaks}
}
\description{
Get the time (in seconds) it takes to decay for each peak
}
