% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_eda_deriv}
\alias{get_eda_deriv}
\title{Electrodermal activity signal derivative}
\usage{
get_eda_deriv(eda)
}
\arguments{
\item{eda}{eda vector}
}
\description{
Finds the first derivatives of the eda signal
}
