% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_half_amp}
\alias{get_half_amp}
\title{Get the halftime of the peaks}
\usage{
get_half_amp(data, i)
}
\arguments{
\item{data}{vector?}

\item{i}{which peak?}
}
\description{
Get the halftime of the peaks
}
