% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_half_rise}
\alias{get_half_rise}
\title{Half rise time}
\usage{
get_half_rise(data, i_apex_with_decay)
}
\arguments{
\item{data}{df with peak info}

\item{i_apex_with_decay}{relevant apices}
}
\description{
Get the time (in seconds) it takes to get to halfway the rise in a peak
}
