% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_max_deriv}
\alias{get_max_deriv}
\title{Maximum derivative}
\usage{
get_max_deriv(data, eda_deriv, sample_rate)
}
\arguments{
\item{data}{df with info on the peaks}

\item{eda_deriv}{derivative of the signal}

\item{sample_rate}{sample rate of the signal}
}
\description{
Get the largest slope before apex, interpolated to seconds
}
