% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_peak_end}
\alias{get_peak_end}
\title{Peak end}
\usage{
get_peak_end(data, max_lookahead)
}
\arguments{
\item{data}{df with peak info}

\item{max_lookahead}{max distance from apex to search for end}
}
\description{
Find the end of the peaks, with some restrictions on the search
}
