% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_peak_end_times}
\alias{get_peak_end_times}
\title{Get the endtimes of the peaks}
\usage{
get_peak_end_times(data)
}
\arguments{
\item{data}{vector?}
}
\description{
Get the endtimes of the peaks
}
