% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_peak_start}
\alias{get_peak_start}
\title{Get the start of the peaks}
\usage{
get_peak_start(data, sample_rate)
}
\arguments{
\item{data}{vector?}

\item{sample_rate}{sample rate of the signal}
}
\description{
Where does a peak start?
}
\details{

}
