% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_peak_start_times}
\alias{get_peak_start_times}
\title{Get the start times of the peaks}
\usage{
get_peak_start_times(data)
}
\arguments{
\item{data}{vector?}
}
\description{
What are the start times of the peaks?
}
