% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_rise_time}
\alias{get_rise_time}
\title{Rise time of peaks}
\usage{
get_rise_time(eda_deriv, apices, sample_rate, start_WT)
}
\arguments{
\item{eda_deriv}{first derivative of signal}

\item{apices}{apex status per measurement (0 or 1)}

\item{sample_rate}{sample rate of the signal}

\item{start_WT}{window within which to look for rise time (in seconds)}
}
\description{
Calculates the rise time of all peaks
}
