% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{get_SCR_width}
\alias{get_SCR_width}
\title{Peak width}
\usage{
get_SCR_width(data, i_apex_with_decay)
}
\arguments{
\item{data}{df with peak info}

\item{i_apex_with_decay}{relevant apices}
}
\description{
Get the width of the peak (in seconds, from halfway the rise until the end)
}
