% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{get_second_derivative}
\alias{get_second_derivative}
\title{Second derivative}
\usage{
get_second_derivative(values)
}
\arguments{
\item{values}{vector of numbers}
}
\description{
Get the second derivative.
}
