% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibi_analysis.R
\name{ibi_analysis}
\alias{ibi_analysis}
\title{Analysis of interbeat interval (IBI)}
\usage{
ibi_analysis(IBI)
}
\arguments{
\item{IBI}{IBI data, component of object read with \code{\link{read_e4}}}
}
\description{
Analysis of interbeat interval (IBI)
}
