% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{max_per_n}
\alias{max_per_n}
\title{Maximum number of output length}
\usage{
max_per_n(values, n, output_length)
}
\arguments{
\item{values}{which values are used}

\item{n}{what is the n}

\item{output_length}{what is the output length}
}
\description{
Maximum number of output length
}
