% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{predict_binary_classifier}
\alias{predict_binary_classifier}
\title{Binary classifiers}
\usage{
predict_binary_classifier(data)
}
\arguments{
\item{data}{features from EDA signal}
}
\description{
Generate classifiers (artifact, no artifact)
}
