% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{predict_multiclass_classifier}
\alias{predict_multiclass_classifier}
\title{Predict multiclass classifier}
\usage{
predict_multiclass_classifier(input)
}
\description{
Predict multiclass classifier
}
