% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_proces_e4.R
\name{read_and_process_e4}
\alias{read_and_process_e4}
\title{Read, process and feature extraction of E4 data}
\usage{
read_and_process_e4(zipfile, tz = Sys.timezone())
}
\arguments{
\item{zipfile}{zip file with e4 data to be read}

\item{tz}{timezone where data were recorded (default system timezone)}
}
\value{
An object with processed data and analyses, object of class 'e4_analysis'.
}
\description{
Reads the raw ZIP file using `read_e4`, 
  performs analyses with `ibi_analysis` and `eda_analysis`.
}
