% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{remove_small_peaks}
\alias{remove_small_peaks}
\title{Remove small peaks}
\usage{
remove_small_peaks(data, thres)
}
\arguments{
\item{data}{vector?}

\item{thres}{which threshold should be used}
}
\description{
Remove small peaks
}
\details{

}
