% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_detection.R
\name{split_in_chunks}
\alias{split_in_chunks}
\title{Compute wavelet features}
\usage{
split_in_chunks(data, rows_per_chunk)
}
\arguments{
\item{data}{vector?}

\item{rows_per_chunk}{number of rows for each window?}
}
\description{
compute wavelet features
}
\details{

}
